pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "unredactor-mcp"
version = "0.1.0"
description = "MCP server for detecting and replacing PDF redaction boxes"
readme = "README.md"
license = "MIT"
authors = [
    { name = "jjohnson" }
]
keywords = ["mcp", "pdf", "redaction", "unredact", "ai", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "PyMuPDF>=1.23.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/Justinandjohnson/unredactor-mcp"
Repository = "https://github.com/Justinandjohnson/unredactor-mcp"
Issues = "https://github.com/Justinandjohnson/unredactor-mcp/issues"

[project.scripts]
unredactor-mcp = "unredactor_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["unredactor_mcp"]

[tool.hatch.build.targets.sdist]
include = ["unredactor_mcp/**"]
